added folded ADC(2) and approximatly folded ADC(3)
added 2 new classes in AdcMatrix:
FoldedAdcMatrix that constructs a folded matrix of any chosen partitioning.
Only the space that is supposed to be folded in has to be given.
Does exact folding if the diagonal block of the chosen folded in space is of order 0, i.e. only diagonal entries.
Implemented new matrix vector product that computes Schur complement, reuses methods of the parent AdcMatrix.

Forwards to another class if the block order is >1.

This class is named ComplementaryBlockInverterNeumann
It applies a Neumann series of the inverse of the shifted complementary space up to any order
This approximates the matvec of the inverted term that arises in folding.

Added slightly adapted Jacobi solver to quickly produce results.
More sophisticated implementation desirable

Adapted workflow in a very adhoc way to quickly produce first results.